refactor: move camera control functions
